% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brap-package.R
\docType{package}
\name{brap-package}
\alias{brap}
\alias{brap-package}
\title{brap: Structured Branching Process Simulation}
\description{
Simulation of renewal-equation structured-population branching process in R for modelling infectious disease outbreaks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/goldingn/brap}
  \item Report bugs at \url{https://github.com/goldingn/brap/issues}
}

}
\author{
\strong{Maintainer}: Nick Golding \email{nick.golding@gmail.com} (\href{https://orcid.org/0000-0001-8916-5570}{ORCID})

}
\keyword{internal}
