% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_ngm.R
\name{calibrate_ngm}
\alias{calibrate_ngm}
\title{Calibrate a Next Generation Matrix}
\usage{
calibrate_ngm(ngm_unscaled, R = 1)
}
\arguments{
\item{ngm_unscaled}{an unscaled next generation matrix}

\item{R}{the target reproduction number of the next generation matrix}
}
\value{
a calibrate next generation matrix
}
\description{
Calibrate a next generation matrix (NGM) to have a specified
reproduction number.
}
\examples{
set.seed(2022-8-30)
n_groups <- 10
contact_matrix <- matrix(
  runif(n_groups ^ 2, 5, 8),
  n_groups,
  n_groups
  )
contact_matrix

ngm <- calibrate_ngm(ngm_unscaled = contact_matrix, R = 2.5)

ngm

}
\author{
Nick Golding
}
