% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_to_matrix.R
\name{cases_to_matrix}
\alias{cases_to_matrix}
\title{Convert a Case Linelist Dataframe to a Matrix}
\usage{
cases_to_matrix(cases, n_groups = max(cases$group))
}
\arguments{
\item{cases}{a case linelist dataframe with three columns as described below}

\item{n_groups}{the total number of groups in the next generation matrix}
}
\value{
a matrix with rows giving dates (all dates between the minimum and
maximum date in 'cases'), and columns giving groups (n_groups, in numerical
order)
}
\description{
Convert a linelist of cases (in the same format as produced by
\code{\link[=sim_cases]{sim_cases()}} into a date-by-group matrix of the number of new infections
on that day in that group. The case linelist must have columns: 'case_id'
(an identifier, in any format); 'group', an integer giving the columns/row
of the next generation matrix to which the case belongs; and
'infection_date' a Date object giving the date of infection of the case.
}
\author{
Nick Golding
}
