% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cases.R
\name{sim_cases}
\alias{sim_cases}
\title{Simulate Fake Linelist Data}
\usage{
sim_cases(n_cases, n_groups, max_days_ago = 14)
}
\arguments{
\item{n_cases}{the number of cases to simulate}

\item{n_groups}{the number of population groups to which cases can belong}

\item{max_days_ago}{the maximum number of days in the past to simulate cases}
}
\value{
a dataframe in the format described in details.
}
\description{
The case linelist must have columns: 'case_id' (an identifier,
in any format); 'group', an integer giving the columns/row of the next
generation matrix to which the case belongs; and 'infection_date' a Date
object giving the date of infection of the case.
}
\author{
Nick Golding
}
