% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_ngm.R
\name{calibrate_ngm}
\alias{calibrate_ngm}
\title{Calibrate a Next Generation Matrix}
\usage{
calibrate_ngm(ngm_unscaled, R = 1)
}
\arguments{
\item{ngm_unscaled}{an unscaled next generation matrix}

\item{R}{the target reproduction number of the next generation smatrix}
}
\value{
a calibrate next generation matrix
}
\description{
Calibrate a next generation matrix (NGM) to have a specified reproduction
number.
}
\author{
Nick Golding
}
