% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Simulate Future Infections}
\usage{
project(ngm, infections_matrix, n_days_project = 10, ...)
}
\arguments{
\item{ngm}{a next generation matrix giving the average rate of new infection
between different population groups}

\item{infections_matrix}{a day-by-group matrix of the number of initial new
infections per day to simulate form}

\item{iterations}{the number of days to simulate beyond the end of
infections_matrix}
}
\value{
a matrix of new infection sper day in each group undr a single
simulation
}
\description{
Simulate daily new infections in each group forward in time with the structured branching
process model.
}
\author{
Nick Golding
}
